requirements.txt
asgiref==3.9.1
autopep8==2.3.2
black==25.1.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dj-database-url==3.0.1
Django==5.2.4
django-jalali==7.4.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
uvicorn==0.35.0
whitenoise==6.9.0
